feat(account): Add routes parameters class validator
